Basicamente ya implementé todo lo que requería xd.
Falta mirar lo del server, creo :c xd
