+++ src/components/Products/ProductsGrid/ProductsGrid.tsx
@@ -3,17 +3,17 @@ import {useFetch} from "../../../hooks";
 import {Product} from "../../../types";
 import {ProductCard} from "../ProductCard/ProductCard";
 import {useConsultInventory} from "../../../hooks/useConsultInventory";
-
+import {CartProvider} from "../../../context/cart.context";
 
 
 export const ProductsGrid = () => {
     const {url} = useConsultInventory();
     const {data, loading, error} = useFetch<Product[]>(url);
 
-    if(error){
+    if (error) {
         return (<> error : {error.message}</>)
     }
-    if(loading){
+    if (loading) {
         return (<> loading : {loading} </>)
     }
 
@@ -21,12 +21,14 @@ export const ProductsGrid = () => {
 
     return (
         <>
-            <div className="products-grid">
-                {data?.map((product) => (
-                    <ProductCard key={product.idInventory} product={product} />
+            <CartProvider>
+                <div className="products-grid">
+                    {data?.map((product) => (
+                        <ProductCard key={product.idInventory} product={product}/>
 
-                ))}
-            </div>
+                    ))}
+                </div>
+            </CartProvider>
         </>
     )
 }
