+++ src/components/Sidebar/Sidebar.css
@@ -8,7 +8,8 @@
     height: 100dvh;
     width: 20em;
     background-color: #1F4F77;
-    z-index: 5;
+    z-index:
+            6;
     transition: transform 0.3s ease-in-out;
 }
 .sidebar.open{
