+++ src/pages/CartPage/CartPage.tsx
@@ -1,14 +1,17 @@
-import {TitleDecorate,CartSummary,CartList} from "../../components";
+import {TitleDecorate, CartSummary, CartList} from "../../components";
 import './CartPage.css'
+import {CartProvider} from "../../context/cart.context";
 
 export const CartPage = () => {
     return (
         <div>
-            <TitleDecorate title={"shopping"} subTitle={"Carrito"}/>
-            <div className="body-shopping">
-                <CartList/>
-                <CartSummary/>
-            </div>
+            <CartProvider>
+                <TitleDecorate title={"shopping"} subTitle={"Carrito"}/>
+                <div className="body-shopping">
+                    <CartList/>
+                    <CartSummary/>
+                </div>
+            </CartProvider>
         </div>
     )
 }
